rename slice to zrange-common for consistency
