+++ src/commands/zrevrange.js
@@ -1,5 +1,5 @@
 import { orderBy } from 'lodash';
-import { slice } from './slice';
+import { slice } from './zrange-command.common';
 
 export function zrevrange(key, s, e) {
   const map = this.data.get(key);
@@ -10,7 +10,9 @@ export function zrevrange(key, s, e) {
   const start = parseInt(s, 10);
   const end = parseInt(e, 10);
 
-  const val = orderBy(Array.from(map.values()), 'score', 'desc').map(it => it.value);
+  const val = orderBy(Array.from(map.values()), 'score', 'desc').map(
+    it => it.value
+  );
 
   return slice(val, start, end);
 }
